HZN-1523: Renaming and minor cleanup
