+++ features/graph/service/src/main/java/org/opennms/netmgt/graph/persistence/converter/CollectionConverter.java
@@ -35,8 +35,8 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.opennms.netmgt.graph.persistence.converter.collection.PersistedCollection;
-import org.opennms.netmgt.graph.persistence.converter.collection.PersistedCollectionEntry;
+import org.opennms.netmgt.graph.persistence.converter.collection.SerializedCollection;
+import org.opennms.netmgt.graph.persistence.converter.collection.SerializedCollectionEntry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -51,11 +51,12 @@ import com.google.gson.JsonSerializationContext;
 import com.google.gson.JsonSerializer;
 
 /**
- * Transforms a Collection from and to a String to be persisted in the database.<br>
- * The Collection may only contain elements of a type that is supported by the ConverterService.<br>
- * The conversation follows the following pattern:<br>
- * Collection <-> PersistedCollection <-> JsonString. The PersistedCollection is the Java representation of the Json object.<br>
- * The elements in the Collection are converted to Strings using the ConverterService.<br>
+ * Transforms a {@link java.util.Collection} from and to a String to be persisted/serialized to the database. <br/>
+ * The {@link java.util.Collection} may only contain elements of a type that is supported by the {@link ConverterService}.<br/>
+ * The conversation follows the following pattern:<br/>
+ * {@link java.util.Collection} <-> {@link SerializedCollection} <-> JSON String.
+ * The {@link SerializedCollection} is the Java representation of the JSON object.<br/>
+ * The elements in the Collection are converted to Strings using the {@link ConverterService}.
  */
 public class CollectionConverter implements Converter<Collection<?>> {
 
@@ -72,24 +73,21 @@ public class CollectionConverter implements Converter<Collection<?>> {
 
     @Override
     public String toStringRepresentation(Collection<?> collection) {
-        List<PersistedCollectionEntry> persistedEntries = collection.stream()
-                .map((entry) -> new PersistedCollectionEntry(entry.getClass(), converterService.toStringRepresentation(entry.getClass(), entry)))
+        final List<SerializedCollectionEntry> persistedEntries = collection.stream()
+                .map((entry) -> new SerializedCollectionEntry(entry.getClass(), converterService.toStringRepresentation(entry.getClass(), entry)))
                 .collect(Collectors.toList());
-        PersistedCollection persistedCollection = new PersistedCollection();
-        persistedCollection.setType(collection.getClass());
-        persistedCollection.setEntries(persistedEntries);
-        return gson.toJson(persistedCollection);
+        SerializedCollection serializedCollection = new SerializedCollection();
+        serializedCollection.setType(collection.getClass());
+        serializedCollection.setEntries(persistedEntries);
+        return gson.toJson(serializedCollection);
     }
 
     @Override
     public Collection<?> toValue(Class<Collection<?>> type, String string) {
+        final SerializedCollection serializedCollection = gson.fromJson(string, SerializedCollection.class);
+        final Collection resurrectedCollection = recreateCollection(serializedCollection.getType());
 
-        Collection resurrectedCollection;
-
-        PersistedCollection persistedCollection = gson.fromJson(string, PersistedCollection.class);
-        resurrectedCollection = recreateCollection(persistedCollection.getType());
-
-        persistedCollection.getEntries()
+        serializedCollection.getEntries()
                 .stream()
                 .map(entry -> converterService.toValue(entry.getType(), entry.getValue()))
                 .forEach(resurrectedCollection::add);
@@ -103,25 +101,22 @@ public class CollectionConverter implements Converter<Collection<?>> {
     }
 
     private Collection<?> recreateCollection(Class<Collection<?>> type) {
-        Collection resurrectedCollection;
         try {
-            resurrectedCollection = type.getConstructor().newInstance();
+            return type.getConstructor().newInstance();
         } catch (NoSuchMethodException | IllegalAccessException | InstantiationException | InvocationTargetException e) {
             LOG.warn("Cannot recreate {}, will use ArrayList instead. Problem: {}", type, e.getMessage());
-            resurrectedCollection = new ArrayList();
+            return new ArrayList();
         }
-        return resurrectedCollection;
     }
 
-    /** We need this adapter so that the class object in PersistedCollection can be (de)serialized. */
+    /** We need this adapter so that the class object in SerializedCollection can be (de)serialized. */
     private final static class ClassAdapter implements JsonSerializer<Class>, JsonDeserializer<Class> {
 
         public JsonElement serialize(Class src, Type typeOfSrc, JsonSerializationContext context) {
             return new JsonPrimitive(src.getName());
         }
 
-        public Class deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-                throws JsonParseException {
+        public Class deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
             try {
                 return Class.forName(json.getAsJsonPrimitive().getAsString());
             } catch (ClassNotFoundException e) {
